Реализовал переключение качеств потока

+++ KinopubApp/Kinopub.UI/Entities/M3u8/M3u8Stream.cs
@@ -7,7 +7,7 @@ namespace Kinopub.UI.Entities.M3u8
     public class M3u8Stream
     {
         public string ProgramId { get; set; }
-        public long Bandwidth { get; set; }
+        public uint Bandwidth { get; set; }
         public int Resolution { get; set; }
         public List<M3u8Audio> AudioTrack { get; set; }
         public List<M3u8Subtitle> SubtitleTrack { get; set; }
